Unionfs: fix oops bug in file_revalidate
Reverses commit "3e9b8b599aa6832e464f86872ca37840090de630" (in unionfs-3.1
tree) which did this:

   ``Unionfs: revalidate files which are open but unlinked

   Need to re-open the file if the starting dentry is on a higher priority
   branch than the currently opened file (which could happen when we ftruncate
   an opened, but unlinked file).''

That fix caused oopses in file_revalidate for some people.  So, I'm
reversing that old fix: it's possible that bug is already fixed thanks to
other code changes.  If not, we'll need to test and fix revalidation of
open/unlinked files differently and carefully.

Signed-off-by: Erez Zadok <ezk@cs.sunysb.edu>